Handle case where $SHELL is unset
Previously the installer would try to use '*' as the shell when the
$SHELL variable was unset which doesn't make sense. This change
introduces logic for trying to infer the user's login shell from other
environment variables, and falls back to 'sh' and prints a warning if
it's not clear what the user's shell is.

Also adds some logic for installing dune when the user's shell is sh
(and other minimal posix shells ash and dash) by modifying ~/.profile.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>