Remove opam's pre-command shell hook (#14)
The `opam init` command adds some code to the user's shell config file
to set up the user's environment to be able to run programs installed by
opam. Firstly the user's PATH variable is prefixed with the bin
directory from the default opam switch. Additionally, opam installs a
pre-command hook into the shell so that before every command is run, the
user's PATH variable is _again_ prefixed with the path to the bin
directory from the current opam switch. This poses a problem for the
dune binary distro, because if the user's current opam switch contains
dune (which is very likely given dune is a dependency of the majority of
opam packages), the instance of dune from the opam switch will take
precedence over the dune instance from the binary distro. As long as
opam's pre-command shell hook is present, opam's bin directory will
always be at the front of PATH when commands are run regardless of any
additional PATH manipulation done by the install script in the user's
shell config, as the former is run before _every_ command, while the
latter is only run once when the shell starts.

This commit changes the install script to add logic to the user's shell
config file to uninstall opam's pre-command shell hook. Modifications to
the PATH performed by opam's setup will still take place when the shell
starts, so programs in the default opam switch will still be runnable
from the user's environment. Users will still also be able to change
their opam environment by running `eval $(opam env)` which will prefix
PATH with the current opam switch's bin directory. The only difference
is that opam's pre-command shell hook will no longer run before each
command is evaluated.

In order for the pre-command hook to be uninstalled, dune's setup must
take place after opam's setup in the user's shell config. For zsh and
fish this works because dune's setup is appended to the same file to
which opam adds its setup logic, so dune's setup is guaranteed to appear
after opam's setup. For bash it's more complicated as `opam init` adds
its setup to ~/.profile by default, but it's possible that users will
move this logic manually to other config files such as ~/.bashrc. This
means that appending dune's logic to the end of (say) ~/.bashrc doesn't
guarantee that it will run after opam's setup, as a user may have
configured their ~/.profile to source their ~/.bashrc and then run the
opam setup. The simplest way to ensure that dune is always initialized
after opam is to detect which bash config file contains opam's setup
logic, and append dune's setup logic to the same file. The install
script is modified to this effect, and an additional prompt is added so
users can override the choice of which config file to modify.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>