When version number is omitted, install latest stable version
When the install is run with no arguments it will install the latest
version of dune with a binary release. Prior to this change this
included unstable versions of dune such as 3.20.0_alpha4. This changes
the behaviour to only consider versions whose version numbers
look stable.

Also adds some additional options to make it possible to test this
change on a fictional set of released versions of dune as at the time of
writing there is only one stable and one unstable version released.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>