Use printf instead of echo to generate shell config
The implementation of the echo command in the /bin/sh that ships with
macos (which is an old version of bash) differs from modern bash
versions as well as other /bin/sh implementations (e.g. busybox) in how
it treats newlines. Specifically when running `echo "a\\nb"`, /bin/sh on
macos would print a literal newline character between the "a" and "b",
whereas other shells would print a newline escape sequence "\n". Due to
this inconsistency the echo command should be avoided if its argument
has a possibility of containing an escaped newline character, which is
the case for the shell config lines generated by the install script.
This change replaces the use of echo with printf, whose behaviour is
more consistent.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>