Infer the user's shell from the parent command
The previous approach of determining the shell from the $SHELL
environment variable almost never worked in practice. This is because
$SHELL is usually not an environment variable but a "shell variable",
and this is not inherited into the install script environment.

Following the approach of `opam init`, this change infers the user's
shell from the parent process of the install script. In the common case
where the script is run from the command line, the user's shell will be
the parent process of the install script. This change also includes
handling for the case where the script is run by `expect`, since that is
how we test the interactive elements of the install script.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>